Finding Donors for CharityML
The client's goal was to construct a model that accurately predicts whether an individual makes more than $50,000.
Understanding an individual's income helped a non-profit better understand how large of a donation to request, or whether or not they should reach out, to begin with. This was done using supervised learning techniques.